Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
pillow==11.1.0
pydantic==2.4.2
email-validator==2.1.1
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
firebase-admin==6.4.0
httpx==0.28.1 